+++ tests/urls.py
@@ -1,4 +1,5 @@
 from django.conf.urls.i18n import i18n_patterns
+from django.contrib import admin
 from django.http import HttpResponse
 from django.urls import include, path
 
@@ -8,6 +9,7 @@ def about(request):
 
 
 urlpatterns = [
+    path("admin/", admin.site.urls),
     path("", include("seo_suite.urls")),  # serves /robots.txt
 ]